+++ client/app/scripts/directives/product-pagination/product-pagination-directive.js
@@ -9,20 +9,10 @@
 	 */
 	 angular.module('gemStoreApp')
 		.directive('paginationCount', function() {
-				return {
-					restrict: 'E',
-					templateUrl: 'scripts/directives/product-pagination/pagination-count.html',
-					controller: 'ProductsPaginationCtrl'
-				};
-		})
-
-		.filter('startFrom', function () {
-			return function (input, start) {
-				if (input) {
-					start = +start;
-					return input.slice(start);
-				}
-				return [];
+			return {
+				restrict: 'E',
+				templateUrl: 'scripts/directives/product-pagination/pagination-count.html',
+				controller: 'ProductsPaginationCtrl'
 			};
-		}); 
+		});
 })();