+++ client/app/scripts/services/products/products-service.js
@@ -8,18 +8,8 @@
      * # productService
      */
 
-    //angular.module('gemStoreApp.productService',['ngResource','Restangular'])
-    angular.module('gemStoreApp.productService',['ngResource'])
-
-        //ngResource factory
-        .factory('productsService', function($resource) {
-            return $resource('/products/:id', {id:'@id'},{
-                'update': { method: 'PUT'}
-            });
+    angular.module('gemStoreApp.productService', [])
+        .factory('productsService', function(Restangular) {
+            return Restangular.service('products');
         });
-
-        //restAngular factory
-        // .factory('productsService', function(Restangular) {
-        //  return Restangular.service('products');
-        // });
 })();